+++ projectile.el
@@ -755,11 +755,8 @@ Set to nil to disable listing submodules contents."
   (cond
    ;; we prefer fd over find
    ;; note that --strip-cwd-prefix is only available in version 8.3.0+
-   ((executable-find "fd")
-    "fd . -0 --type f --color=never --strip-cwd-prefix")
-   ;; fd's executable is named fdfind is some Linux distros (e.g. Ubuntu)
-   ((executable-find "fdfind")
-    "fdfind . -0 --type f --color=never --strip-cwd-prefix")
+   (projectile-fd-executable
+    (format "%s . -0 --type f --color=never --strip-cwd-prefix" projectile-fd-executable))
    ;; with find we have to be careful to strip the ./ from the paths
    ;; see https://stackoverflow.com/questions/2596462/how-to-strip-leading-in-unix-find
    (t "find . -type f | cut -c3- | tr '\\n' '\\0'"))